Fix broken test in test_queue.js
The queue.add operation was setting the job id in the outer scope
for the process() function to compare it against the processed job's
id later in the test. This was relying on the .then() clause being
executed before the handler - which was causing a race condition.
If the .then() clause could not set the variable in time, it would
remain undefined, therefore making the id equality assertion fail.
This would end the test early with a call to done() therefore starting
the following tests. It would then take another 100ms before
setTimeout() would call jobDone(), which would trigger 'complete',
which would in turn close `anotherQueue` - which in the meanwhile
would be processing jobs sent to the queue. This would cascade down
to the following tests since they all use the same queue name.
With this fix the test is finished only when the first queue has
finished executing and the second queue is closed.