Fix bug in Queue.prototype.close
It returns a promise which should be resolved when
both the client and the blocking client's stream had triggered the
'end' event.
It was incorrectly setting two listeners on the non-blocking client's
stream, instead of one on each. Interestingly this would still work
as this would mean that the listener would be executed twice.
This commit fixes this subtle bug and adds tests for the fix.