Only count locally processed jobs when locally pausing queue.
If we count how many jobs are active globally, and then wait for
that many jobs to complete locally, this promise will never be
resolved, because some other process might be working on some of
those active jobs that were counted.

Since we are really interested in how many jobs are in progress in
this particular process, we should use count of active local jobs,
rather than a count of all active jobs.