Add defaultJobOptions to Queue
I'd like to pass `removeOnComplete` and `removeOnFail` only once,
instead of every time I run `.add()` on a queue. This patch adds a new
option to queues called `defaultJobOptions` which, as the name suggests,
sets the default job options.

Also adds two tests verifying that the option is passed through
correctly and that it works with `removeOnComplete`. I wasn't sure
whether to more tests were necessary, if so, please just let me know and
I'll work on it.